refactor : 장르 컴포넌트 분리 완료
