+++ src/app/main/page.tsx
@@ -1,16 +1,3 @@
-/* eslint-disable import/order */
-/* eslint-disable import/no-named-as-default */
-/* eslint-disable no-console */
-/* eslint-disable no-empty */
-/* eslint-disable no-unused-vars */
-/* eslint-disable import/no-extraneous-dependencies */
-/* eslint-disable @next/next/no-img-element */
-/* eslint-disable jsx-a11y/alt-text */
-/* eslint-disable react-hooks/rules-of-hooks */
-/* eslint-disable max-len */
-/* eslint-disable import/extensions */
-/* eslint-disable import/no-unresolved */
-
 'use client';
 
 import { useState } from 'react';
@@ -20,11 +7,12 @@ import { createTheme } from '@mui/material/styles';
 import { ThemeProvider } from '@emotion/react';
 import ForwardIcon from '@mui/icons-material/ArrowForwardIos';
 import BackIcon from '@mui/icons-material/ArrowBackIosNew';
-import PlayCircleIcon from '@mui/icons-material/PlayCircle';
-import AlbumCoverUser from '../components/AlbumCover/AlbumCoverUser';
-import AlbumCoverSystem from '../components/AlbumCover/AlbumCoverSystem';
-import PopularMusicComponent from './popularMusicComponent';
-import PopularTagComponent from './popularTagComponent';
+
+import AlbumCoverUser from '../components/AlbumCover/AlbumCoverUser.tsx';
+import AlbumCoverSystem from '../components/AlbumCover/AlbumCoverSystem.tsx';
+import PopularMusicComponent from './popularMusicComponent.tsx';
+import PopularTagComponent from './popularTagComponent.tsx';
+import GenreMusicComponent from './genreMusicComponent.tsx';
 
 const theme = createTheme({
   palette: {
@@ -39,79 +27,10 @@ const theme = createTheme({
   },
 });
 
-// 장르별 음악 컴포넌트의 props
-type GenreMusicProps = {
-  genre: string;
-  bgColor: number;
-  musicImage1: string;
-  musicImage2: string;
-  musicImage3: string;
-  musicTitle1: string;
-  musicTitle2: string;
-  musicTitle3: string;
-};
-// 장르별 음악 컴포넌트
-function GenreMusic({
-  genre,
-  bgColor,
-  musicImage1,
-  musicImage2,
-  musicImage3,
-  musicTitle1,
-  musicTitle2,
-  musicTitle3,
-}: GenreMusicProps) {
-  return (
-    <div
-      className={`flex flex-col items-center w-72 h-96 rounded-2xl bg-genre-${bgColor} my-12 m-4 mx-8`}
-    >
-      {/* 위에 단색 배경바 */}
-      <div
-        className={`flex flex-row items-center w-full h-20 p-8 rounded-t-2xl bg-nv-${bgColor}`}
-      >
-        <p className="w-11/12 text-2xl text-black font-bold">{genre}</p>
-        <IconButton>
-          <PlayCircleIcon style={{ fontSize: 50, opacity: 0.7 }} />
-        </IconButton>
-      </div>
-
-      {/* 음악 정보 */}
-      {/* 1번째 음악 */}
-      <div className="flex flex-row items-center justify-start m-2 cursor-pointer hover-bg-opacity">
-        <img className="w-1/4 m-2" src={musicImage1} />
-        <p className="text-lg text-black">{musicTitle1}</p>
-      </div>
-
-      {/* 2번째 음악 */}
-      <div className="flex flex-row items-center justify-start m-2 cursor-pointer hover-bg-opacity">
-        <img className="w-1/4 m-2" src={musicImage2} />
-        <p className="text-lg text-black">{musicTitle2}</p>
-      </div>
-
-      {/* 3번째 음악 */}
-      <div className="flex flex-row items-center justify-start m-2 cursor-pointer hover-bg-opacity">
-        <img className="w-1/4 m-2" src={musicImage3} />
-        <p className="text-lg text-black">{musicTitle3}</p>
-      </div>
-    </div>
-  );
-}
-
 // 메인 페이지 컴포넌트
 function Page() {
-  const [genrePageIndex, setgenrePageIndex] = useState(0); // 장르별 음악 페이지 인덱스
-  const [otherPlaylistPageIndex, setOtherPlaylistPageIndex] = useState(0); // 다른 유저가 선택한 플레이리스트 페이지 인덱스
+  const [otherPlaylistPageIndex, setOtherPlaylistPageIndex] = useState(0);
   const [systemPlaylistPageIndex, setSystemPlaylistPageIndex] = useState(0); // 시스템 플레이리스트 페이지 인덱스
-
-  const handleGenrePageForwardClick = () => {
-    setgenrePageIndex(genrePageIndex + 1);
-  };
-
-  const handleGenrePageBackwardClick = () => {
-    if (genrePageIndex > 0) {
-      setgenrePageIndex(genrePageIndex - 1);
-    }
-  };
 
   const handleOtherPlaylistPageForwardClick = () => {
     setOtherPlaylistPageIndex(otherPlaylistPageIndex + 1);
@@ -153,56 +72,7 @@ function Page() {
           {/* 장르별 음악 */}
           <h1 className="">장르별 음악</h1>
           <div className="flex flex-row justify-center items-center w-full h-auto bg-gray-650 rounded-2xl">
-            <IconButton className="w-4" onClick={handleGenrePageBackwardClick}>
-              {genrePageIndex !== 0 && (
-                <BackIcon color="primary" fontSize="large" />
-              )}
-            </IconButton>
-            <div className="w-11/12 h-full flex flex-row items-center justify-start">
-              <GenreMusic
-                genre="아이돌"
-                bgColor={100}
-                musicImage1="https://i.ibb.co/4d0pj5j/dynamite.webp"
-                musicImage2="https://i.ibb.co/ZcJxmcZ/supershy.webp"
-                musicImage3="https://i.ibb.co/DGWrD6M/image.jpg"
-                musicTitle1="Dynamite"
-                musicTitle2="Super Shy"
-                musicTitle3="불꽃놀이"
-              />
-              <GenreMusic
-                genre="인디"
-                bgColor={200}
-                musicImage1="https://i.ibb.co/4d0pj5j/dynamite.webp"
-                musicImage2="https://i.ibb.co/ZcJxmcZ/supershy.webp"
-                musicImage3="https://i.ibb.co/DGWrD6M/image.jpg"
-                musicTitle1="Dynamite"
-                musicTitle2="Super Shy"
-                musicTitle3="불꽃놀이"
-              />
-              <GenreMusic
-                genre="발라드"
-                bgColor={300}
-                musicImage1="https://i.ibb.co/4d0pj5j/dynamite.webp"
-                musicImage2="https://i.ibb.co/ZcJxmcZ/supershy.webp"
-                musicImage3="https://i.ibb.co/DGWrD6M/image.jpg"
-                musicTitle1="Dynamite"
-                musicTitle2="Super Shy"
-                musicTitle3="불꽃놀이"
-              />
-              <GenreMusic
-                genre="힙합"
-                bgColor={400}
-                musicImage1="https://i.ibb.co/4d0pj5j/dynamite.webp"
-                musicImage2="https://i.ibb.co/ZcJxmcZ/supershy.webp"
-                musicImage3="https://i.ibb.co/DGWrD6M/image.jpg"
-                musicTitle1="Dynamite"
-                musicTitle2="Super Shy"
-                musicTitle3="불꽃놀이"
-              />
-            </div>
-            <IconButton onClick={handleGenrePageForwardClick}>
-              <ForwardIcon color="primary" fontSize="large" />
-            </IconButton>
+            <GenreMusicComponent />
           </div>
 
           {/* 다른 유저가 선택한 플레이리스트 */}
