+++ src/types/genre.ts
@@ -11,3 +11,15 @@ export interface Genre {
   genre_name: string;
   state: boolean;
 }
+
+// 메인 페이지에서 장르를 선택하고 선택된 장르의 props (진우)
+export interface GenreMusicProps {
+  genre: string;
+  bgColor: number;
+  musicImage1: string;
+  musicImage2: string;
+  musicImage3: string;
+  musicTitle1: string;
+  musicTitle2: string;
+  musicTitle3: string;
+}
